Add Shape enum to avatar component, refactor UITextSize to styles/typography.rs
