+++ crates/ui2/src/styles.rs
@@ -1,2 +1,4 @@
 mod elevation;
+mod typography;
 pub use elevation::*;
+pub use typography::*;
